#14 refactor after review
