--- assignment/phase1/code/logics/__init__.py
@@ -1,4 +1,0 @@
-from .database import init_db
-
-def create_app():
-    init_db()
