+++ assignment/phase1/code/process/__init__.py
@@ -1,0 +1,4 @@
+from .database.database import init_db
+
+def create_app():
+    init_db()
