+++ assignment/phase1/code/process/logics/android_data_collector.py
