+++ assignment/phase1/code/process/models/app.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column, String, Float, Integer, Text
-from ..database import Base
+from ..database.database import Base
 
 class AppData(Base):
     __abstract__ = True
